fix open file

+++ sources/open/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: gbertin <gbertin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/02 14:21:45 by gbertin           #+#    #+#             */
-/*   Updated: 2022/10/05 11:48:06 by gbertin          ###   ########.fr       */
+/*   Updated: 2022/10/11 14:16:08 by gbertin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,13 +29,21 @@ static void	print_in_file(int fd, char *limiter, t_minishell *ms)
 	free(append);
 }
 
+static void heredoc_child(int fd, char *limiter, t_minishell *ms)
+{
+	signal(SIGINT, sigint_heredoc);
+	print_in_file(fd, limiter, ms);
+	close(fd);
+	exit_child(0, ms);
+}
+
 int	heredoc(char *limiter, t_minishell *ms)
 {
 	int		fd;
 	int		pid;
 	char	*tmp_file;
 
-	tmp_file = ".tmp";
+	tmp_file = ft_strjoin("/tmp/", ft_itoa(*limiter, ms), ms);
 	if (!tmp_file)
 		return (-1);
 	fd = open(tmp_file, O_TRUNC | O_CREAT | O_RDWR, 0644);
@@ -45,14 +53,12 @@ int	heredoc(char *limiter, t_minishell *ms)
 	if (pid < 0)
 		return (-1);
 	if (pid == 0)
-	{
-		signal(SIGINT, sigint_heredoc);
-		print_in_file(fd, limiter, ms);
-		exit_child(0, ms);
-	}
+		heredoc_child(fd, limiter, ms);
 	waitpid(pid, NULL, 0);
 	close(fd);
 	fd = open(tmp_file, O_RDONLY);
+	if (fd == -1)
+		return (-1);
 	unlink(tmp_file);
 	return (fd);
 }
